feat: add provider to app layer
