+++ src/app/App.tsx
@@ -1,11 +1,11 @@
-import { BrowserRouter as Router } from "react-router-dom"
 import Header from "../widgets/ui/Header.tsx"
 import Footer from "../widgets/ui/Footer.tsx"
 import PostsManagerPage from "../pages/PostsManagerPage.tsx"
+import { Provider } from "./providers/provider.tsx"
 
 const App = () => {
   return (
-    <Router>
+    <Provider>
       <div className="flex flex-col min-h-screen">
         <Header />
         <main className="flex-grow container mx-auto px-4 py-8">
@@ -13,7 +13,7 @@ const App = () => {
         </main>
         <Footer />
       </div>
-    </Router>
+    </Provider>
   )
 }
