+++ src/app/providers/provider.tsx
@@ -1,0 +1,20 @@
+import { BrowserRouter as Router } from "react-router-dom"
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: false,
+      refetchOnWindowFocus: false,
+      staleTime: 1000 * 60 * 3,
+    },
+  },
+})
+
+export const Provider = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <QueryClientProvider client={queryClient}>
+      <Router>{children}</Router>
+    </QueryClientProvider>
+  )
+}
